chunk code rework (again) + half-implemented collision detection
chunks are now clone-on-write, and will orphan any readers that still have the chunk "locked" when it writted to. there is also a write queue, and changes to the chunk are deferred until the end up the update, when each chunk that has been potentially modified is visited and the actual writes and orphaning take place.

collision is still very much a WIP. collisions agains positive-facing block faces seem to be working correctly, but collisions against negative-facing ones are all screwed up.

a bunch of other misc. changes as well.