(mostly) fixed terrain collision detection
the previous issues with collision code were caused by the maximum bound of the detection loops. when pressed against the negative-facing side of a block, that block would be included in the loop. the fix was to shrink the range by 1 if the floor of the maximum was an integer.

note that there is still an issue with this method of collision resolution; a literal corner case if you will :p. when you move directly into a corner, none of the axis projections collide with anything, so no displacement is set, even though the combined movment *would* collide with a block, and you end up jittering an and out of the block.