world: started to migrate to new representation
the world now stores a map of (x, z) -> chunk column, and each column has a map of y -> chunk. this allows some data like heightmaps and sky light nodes to be stored naturally in the column itself and not in a parallel map.

additionally, chunk loading and unloading now uses locks instead of concurrent maps.

work still needs to be done here, like storing modified heightmap information, using the column's heightmap instead of the heightmap cache's version, and more!